Refactors and about to get some updates 🐬
