+++ lib/src/televerse/models/models.dart
@@ -2,6 +2,10 @@ library televerse.models;
 
 import 'dart:io' as io;
 import 'package:televerse/src/types/input_file_type.dart';
+import 'package:televerse/televerse.dart';
 
 part 'chat_id.dart';
 part 'input_file.dart';
+part 'webhook_exception.dart';
+part 'http_exception.dart';
+part 'long_polling_exception.dart';
