+++ lib/src/televerse/televerse.dart
@@ -1,1 +1,202 @@
-export './models/models.dart';
+part of televerse;
+
+class Televerse {
+  final String token;
+
+  late Fetcher fetcher;
+
+  Televerse(
+    this.token, {
+    Fetcher? fetcher,
+  }) {
+    this.fetcher = fetcher ?? LongPolling(this);
+  }
+
+  final String _baseUrl = "https://api.telegram.org/bot";
+  Uri _buildUri(String method, [Map<String, dynamic>? params]) {
+    return Uri.https(_baseUrl, "$token/$method", params);
+  }
+
+  ///
+  void onUpdate(Update update) {}
+
+  /// Start polling for updates.
+  Future<void> start() async {
+    await fetcher.start();
+    fetcher.onUpdate().listen((update) {
+      if (update.message != null) {
+        onUpdate(update);
+      }
+    });
+  }
+
+  /// Use this method to receive incoming updates using long polling.
+  /// An Array of Update objects is returned.
+  ///
+  /// See more at https://core.telegram.org/bots/api#getupdates
+  Future<List<Update>> getUpdates({
+    int? offset,
+    int? limit,
+    int? timeout,
+    List<String>? allowedUpdates,
+  }) async {
+    Map<String, dynamic> params = {
+      "offset": offset,
+      "limit": limit,
+      "timeout": timeout,
+      "allowed_updates": jsonEncode(allowedUpdates),
+    };
+    Uri uri = _buildUri("getUpdates", params);
+
+    List response = await HttpClient.get(uri);
+    return (response)
+        .map((e) => Update.fromJson(e as Map<String, dynamic>))
+        .toList();
+  }
+
+  /// Use this method to specify a URL and receive incoming updates via an outgoing webhook.
+  /// Whenever there is an update for the bot, we will send an HTTPS POST request to the specified URL,
+  /// containing a JSON-serialized Update. In case of an unsuccessful request, we will give up after a
+  /// reasonable amount of attempts. Returns True on success.
+  /// If you'd like to make sure that the webhook was set by you, you can specify secret data in the
+  /// parameter secret_token. If specified, the request will contain a header “X-Telegram-Bot-Api-Secret-Token”
+  /// with the secret token as content.
+  ///
+  /// See more at https://core.telegram.org/bots/api#setwebhook
+  Future<bool> setWebhook({
+    required String url,
+    io.File? certificate,
+    String? ipAddress,
+    int? maxConnections,
+    List<String>? allowedUpdates,
+    bool? dropPendingUpdates,
+    String? secretToken,
+  }) async {
+    Map<String, dynamic> params = {
+      "url": url,
+      "ip_address": ipAddress,
+      "max_connections": maxConnections,
+      "allowed_updates": jsonEncode(allowedUpdates),
+      "drop_pending_updates": dropPendingUpdates,
+      "secret_token": secretToken,
+    };
+    if (certificate != null) {
+      List<MultipartFile> files = [];
+      files.add(MultipartFile(
+        'certificate',
+        certificate.openRead(),
+        certificate.lengthSync(),
+        filename: certificate.path.split("/").last,
+      ));
+      Uri uri = _buildUri("setWebhook");
+      return await HttpClient.multipartPost(uri, files, params);
+    } else {
+      Uri uri = _buildUri("setWebhook");
+      return await HttpClient.post(uri, params);
+    }
+  }
+
+  /// Use this method to remove webhook integration if you decide to switch back to getUpdates.
+  /// Returns True on success.
+  ///
+  /// See more at https://core.telegram.org/bots/api#deletewebhook
+  Future<bool> deleteWebhook({bool? dropPendingUpdates}) async {
+    Map<String, dynamic> params = {
+      "drop_pending_updates": dropPendingUpdates,
+    };
+    Uri uri = _buildUri("deleteWebhook", params);
+
+    bool response = await HttpClient.get(uri);
+    return response;
+  }
+
+  /// Use this method to get current webhook status. Requires no parameters. On success, returns a WebhookInfo object.
+  /// If the bot is using getUpdates, will return an object with the url field empty.
+  ///
+  /// See more at https://core.telegram.org/bots/api#getwebhookinfo
+  Future<WebhookInfo> getWebhookInfo() async {
+    Uri uri = _buildUri("getWebhookInfo");
+
+    Map<String, dynamic> response = await HttpClient.get(uri);
+    return WebhookInfo.fromJson(response);
+  }
+
+  /// A simple method for testing your bot's authentication token. Requires no parameters.
+  /// Returns basic information about the bot in form of a User object.
+  ///
+  /// See more at https://core.telegram.org/bots/api#getme
+  Future<User> getMe() async {
+    Uri uri = _buildUri("getMe");
+
+    Map<String, dynamic> response = await HttpClient.get(uri);
+    return User.fromJson(response);
+  }
+
+  /// Use this method to log out from the cloud Bot API server before launching the bot locally. You must log out the bot before running it locally, otherwise there is no guarantee that the bot will receive updates. After a successful call, you can immediately log in on a local server, but will not be able to log in back to the cloud Bot API server for 10 minutes. Returns True on success. Requires no parameters.
+  /// Note: This will affect all running bots.
+  ///
+  /// See more at https://core.telegram.org/bots/api#logout
+  Future<bool> logOut() async {
+    Uri uri = _buildUri("logOut");
+
+    bool response = await HttpClient.get(uri);
+    return response;
+  }
+
+  /// Use this method to close the bot instance before moving it from one local server to another. You need to delete the webhook before calling this method to ensure that the bot isn't launched again after server restart. The method will return error 429 in the first 10 minutes after the bot is launched. Returns True on success. Requires no parameters.
+  ///
+  /// See more at https://core.telegram.org/bots/api#close
+  Future<bool> close() async {
+    Uri uri = _buildUri("close");
+
+    bool response = await HttpClient.get(uri);
+    return response;
+  }
+
+  /// Use this method to send text messages. On success, the sent [MessageContext] is returned.
+  ///
+  /// Required parameters:
+  /// - [chatId] - Chat ID can either be [ChatID] or [ChannelID] or [SupergroupID]
+  ///
+  /// - [text] - Text of the message to be sent
+  ///
+  /// You can optionally pass the more parameters as described in the official documentation.
+  /// See more at https://core.telegram.org/bots/api#sendmessage
+  ///
+  /// **Example:**
+  /// ```dart
+  /// /// If you're using [ChatID] specify the chat id as an integer
+  /// bot.sendMessage(ChatID(123456789), "Hello World!");
+  ///
+  /// /// If you're using [ChannelID] or [SupergroupID] specify the chat id as a string
+  /// /// Both [ChannelID] and [SupergroupID] are treated the same way
+  /// bot.sendMessage(ChannelID("@myChannel"), "Hello World!");
+  /// ```
+  ///
+  /// On success, the sent [MessageContext] is which can be used to reply to the message.
+  Future<MessageContext> sendMessage({
+    required ID chatId,
+    required String text,
+    String? parseMode,
+    bool? disableWebPagePreview,
+    bool? disableNotification,
+    int? replyToMessageId,
+    bool? allowSendingWithoutReply,
+    ReplyMarkup? replyMarkup,
+  }) async {
+    Map<String, dynamic> params = {
+      "chat_id": chatId,
+      "text": text,
+      "parse_mode": parseMode,
+      "disable_web_page_preview": disableWebPagePreview,
+      "disable_notification": disableNotification,
+      "reply_to_message_id": replyToMessageId,
+      "allow_sending_without_reply": allowSendingWithoutReply,
+      "reply_markup": jsonEncode(replyMarkup?.toJson()),
+    };
+    Uri uri = _buildUri("sendMessage", params);
+
+    Map<String, dynamic> response = await HttpClient.get(uri);
+    return MessageContext(this, Message.fromJson(response));
+  }
+}
