+++ lib/telegram.dart
@@ -1,0 +1,3 @@
+library telegram;
+
+export 'src/telegram/models/models.dart';
